refactor: Participant 추상클래스로 추출
