+++ src/main/kotlin/blackjackgame/model/player/Players.kt
@@ -2,7 +2,7 @@ package blackjackgame.model.player
 
 import blackjackgame.model.card.Deck
 
-class Players(val players: List<Player>) : List<Player> by players {
+class Players(val players: List<Participant>) : List<Participant> by players {
 
     fun drawInitCards(deck: Deck) {
         players.forEach {
